# apps/credential-manager/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/credentials/{credentialRecordId}": {
      "get": {
        "operationId": "CredentialsController_get",
        "summary": "Fetch a credential",
        "description": "This call provides a credential for a given tenant",
        "parameters": [
          {
            "name": "credentialRecordId",
            "required": true,
            "in": "path",
            "description": "The credential record ID to retrieve",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential fetched successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Internal server error": {
                    "value": {
                      "statusCode": 500,
                      "message": "Internal server error",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": ["Credentials"]
      }
    },
...
    "/v1/credential-offers": {
      "get": {
        "operationId": "CredentialOffersController_find",
        "summary": "Fetch a list of credential offers",
        "description": "This call provides a list of credential offers for a given tenant",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential offers fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offers fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential offers fetched successfully",
                      "data": [{ "id": "71b784a3" }]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Something went wrong": {
                    "value": {
                      "statusCode": 500,
                      "message": "Something went wrong",
                      "error": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Offers"]
      },
      "post": {
        "operationId": "CredentialOffersController_offer",
        "summary": "Create a credential offer",
        "description": "This call creates a credential offer for a given connection ID and credential definition ID",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OfferPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential offer created successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer created successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential offer created successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential offer not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  },
                  "Credential definition not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential definition not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Something went wrong": {
                    "value": {
                      "statusCode": 500,
                      "message": "Something went wrong",
                      "error": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Offers"]
      }
    },
    "/v1/credential-offers/{credentialOfferId}": {
      "get": {
        "operationId": "CredentialOffersController_getById",
        "summary": "Fetch a credential offer by ID",
        "description": "This call provides a credential offer for a given ID",
        "parameters": [
          {
            "name": "credentialOfferId",
            "required": true,
            "in": "path",
            "description": "The credential offer ID to retrieve",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential offer fetched successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer fetched successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential offer fetched successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential offer not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Something went wrong": {
                    "value": {
                      "statusCode": 500,
                      "message": "Something went wrong",
                      "error": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Offers"]
      }
    },
    "/v1/credential-offers/self": {
      "post": {
        "operationId": "CredentialOffersController_offerToSelf",
        "summary": "Create a credential offer to self",
        "description": "This call creates a credential offer for a given credential definition ID",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OfferPayloadSelf" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential offer created successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer created successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential offer created successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential offer not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  },
                  "Credential definition not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential definition not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Something went wrong": {
                    "value": {
                      "statusCode": 500,
                      "message": "Something went wrong",
                      "error": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Offers"]
      }
    },
    "/v1/credential-offers/{credentialOfferId}/accept": {
      "post": {
        "operationId": "CredentialOffersController_acceptOffer",
        "summary": "Accept a credential offer",
        "description": "This call accepts a credential offer for a given ID",
        "parameters": [
          {
            "name": "credentialOfferId",
            "required": true,
            "in": "path",
            "description": "The credential offer ID to retrieve",
            "schema": { "format": "string", "type": "string" }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "description": "Specifies the tenant ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Credential offer accepted successfully",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer accepted successfully": {
                    "value": {
                      "statusCode": 200,
                      "message": "Credential offer accepted successfully",
                      "data": { "id": "71b784a3" }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Credential offer not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Credential offer not found",
                      "data": null
                    }
                  },
                  "Tenant not found": {
                    "value": {
                      "statusCode": 404,
                      "message": "Tenant not found",
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          },
          "500": {
            "description": "Something went wrong",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "Something went wrong": {
                    "value": {
                      "statusCode": 500,
                      "message": "Something went wrong",
                      "error": "Internal Server Error"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["Credential Offers"]
      }
    },
...
  },
  "info": {
    "title": "Gaia-X OCM Credential Manager",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Attribute": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "value": { "type": "string" },
          "mimeType": { "type": "string" }
        },
        "required": ["name", "value"]
      },
      "OfferPayload": {
        "type": "object",
        "properties": {
          "connectionId": { "type": "string" },
          "credentialDefinitionId": { "type": "string" },
          "attributes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Attribute" }
          },
          "revocationRegistryDefinitionId": { "type": "string" }
        },
        "required": [
          "connectionId",
          "credentialDefinitionId",
          "attributes",
          "revocationRegistryDefinitionId"
        ]
      },
      "OfferPayloadSelf": {
        "type": "object",
        "properties": {
          "credentialDefinitionId": { "type": "string" },
          "attributes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Attribute" }
          },
          "revocationRegistryDefinitionId": { "type": "string" }
        },
        "required": [
          "credentialDefinitionId",
          "attributes",
          "revocationRegistryDefinitionId"
        ]
      },
...
    }
  }
}